% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{calibration_output_vars}
\alias{calibration_output_vars}
\title{Define output variables for calibration runs}
\usage{
calibration_output_vars()
}
\arguments{
\item{lower_age}{Vector of lower bounds for age groups}

\item{upper_age}{Vector of upper bounds for age groups}
}
\description{
Prevalence, incidence, severe incidence and proportion in age group
}
