% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{life_expectancy}
\alias{life_expectancy}
\title{Join life expectancy}
\usage{
life_expectancy(x, le)
}
\arguments{
\item{x}{data}

\item{le}{Life expectancy table}
}
\description{
Join life expectancy
}
