% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{model_output_to_long}
\alias{model_output_to_long}
\title{Raw output to long}
\usage{
model_output_to_long(x, ...)
}
\arguments{
\item{x}{Model output}

\item{...}{Additional columns to select (for example run names)}
}
\value{
Long output
}
\description{
Converts wide (by age groups) to long
}
