% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{model_run}
\alias{model_run}
\title{Run the model}
\usage{
model_run(pfpr, rtss_coverage, draw = 0, season = "perennial")
}
\arguments{
\item{pfpr}{PRevalence band 2-10}

\item{rtss_coverage}{Vaccine coverage}

\item{draw}{Parameter draw}

\item{season}{Seasonality}
}
\description{
Run the model
}
