% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{output_vars}
\alias{output_vars}
\title{Define output variables}
\usage{
output_vars(lower_age = 0:99, upper_age = 1:100)
}
\arguments{
\item{lower_age}{Vector of lower bounds for age groups}

\item{upper_age}{Vector of upper bounds for age groups}
}
\description{
Prevalence, incidence, severe incidence and propotion in age group
}
