% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{seasonality}
\alias{seasonality}
\title{Define seasonal profiles}
\usage{
seasonality(type)
}
\arguments{
\item{type}{Type of profile, this can either be "perennial" or "seasonal". The perennial
is flat, the seasonal profile is lifted from the rainfall fit for Bamako in Mali.}
}
\description{
Define seasonal profiles
}
