% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{treatment_option}
\alias{treatment_option}
\title{Treatment options}
\usage{
treatment_option(act_coverage = 0.2647059, non_act_coverage = 0.2647059)
}
\arguments{
\item{act_coverage}{ACT coverage}

\item{non_act_coverage}{non ACT coverage}
}
\description{
Following Penny et al:
In the Imperial model this was implemented as 52\% access to ACT or non-ACT treatment (26\% each), with efficacies of 95\% and 75\% respectively.
}
