% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{vectors}
\alias{vectors}
\title{Vector species proportions}
\usage{
vectors(prop_gamb_ss = 0.5, prop_fun = 0.25, prop_arab = 0.25)
}
\arguments{
\item{prop_gamb_ss}{Proportion gambiae ss}

\item{prop_fun}{Proportion funestus}

\item{prop_arab}{Proportion arabiensis}
}
\description{
Vector species proportions
}
