% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{year_summary}
\alias{year_summary}
\title{Isolate annual summary}
\usage{
year_summary(x, baseline_year = 2000, max_year = 2050)
}
\arguments{
\item{x}{Model output}

\item{baseline_year}{Start year}

\item{max_year}{End year}
}
\value{
Annual output
}
\description{
To ensure compatible run options as GTS modelling we return monthly output. However,
to reduce storage and processing requirements we immediately trim to annual output.
}
